data-prep/requirements.txt
# Data processing and metadata
pillow
colorgram.py
pandas
numpy
pytesseract
opencv-python
matplotlib
scikit-learn
jsonschema
tesseract

# CLIP model and PyTorch
jinja2
torch
torchvision
ftfy
regex
tqdm
git+https://github.com/openai/CLIP.git



# Firebase Integration
firebase-admin